backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[bcrypt]
python-dotenv
python-jose[cryptography]
pydantic[email]
bcrypt==3.2.2
python-socketio[asyncio]